feat(dedicated.cluster): move bm cluster component to dedicated app
ref: MANAGER-12269
