+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/cluster.html
