+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/dashboard/dashboard.controller.js
