+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/dashboard/general-information/general-information.controller.js
@@ -1,4 +1,4 @@
-export default class BmClusterComponentsGeneralInformationController {
+export default class DedicatedClusterGeneralInformationController {
   /* @ngInject */
   constructor($state) {
     this.$state = $state;
