+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/dashboard/general-information/general-information.html
