+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/dashboard/general-information/index.js
@@ -3,13 +3,12 @@ import angular from 'angular';
 import '@ovh-ux/ui-kit';
 import 'angular-translate';
 
-import ngAtInternet from '@ovh-ux/ng-at-internet';
 import component from './general-information.component';
 
-const moduleName = 'ovhManagerBmClusterComponentsDashboardGeneralInformation';
+const moduleName = 'ovhManagerClusterComponentsDashboardGeneralInformation';
 
 angular
-  .module(moduleName, [ngAtInternet, 'oui', 'pascalprecht.translate'])
+  .module(moduleName, ['oui', 'pascalprecht.translate'])
   .component('clusterGeneralInformation', component)
   .run(/* @ngTranslationsInject:json ./translations */);
