+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/dashboard/general-information/translations/Messages_fr_FR.json
@@ -14,12 +14,12 @@
   "dedicated_cluster_dashboard_region_us-east-vin": "North America (US - East - Vinthill)",
   "dedicated_cluster_dashboard_region_us-west-hil": "North America (US - West - Hillsboro)",
   "dedicated_cluster_dashboard_region_ca-east-bhs": "North America (Canada - East - Beauharnois)",
-  "dedicated_cluster_dashboard_region_ap-southeast-sgp": "Asia Pacific (Singapore -Singapore)",
+  "dedicated_cluster_dashboard_region_ap-southeast-sgp": "Asia Pacific (Singapore - Singapore)",
   "dedicated_cluster_dashboard_region_ap-southeast-syd": "Asia Pacific (Australia - Sydney)",
   "dedicated_cluster_dashboard_region_eu-west-rbx-snc": "Europe (France - Roubaix) (SNC)",
   "dedicated_cluster_dashboard_region_eu-west-sbg-snc": "Europe (France - Strasbourg) (SNC)",
-  "dedicated_cluster_dashboard_region_ca-east-tor": "North America (Canada - East  - Toronto)",
-  "dedicated_cluster_dashboard_region_ap-south-mum": "Asia Pacific (India – Mumbai)",
+  "dedicated_cluster_dashboard_region_ca-east-tor": "North America (Canada - East - Toronto)",
+  "dedicated_cluster_dashboard_region_ap-south-mum": "Asia Pacific (India - Mumbai)",
   "dedicated_cluster_dashboard_region_labeu-west-1-preprod": "LABEU (France - Croix) (PREPROD)",
   "dedicated_cluster_dashboard_region_labeu-west-1-dev-1": "LABEU (France - Croix) (DEV-1)",
   "dedicated_cluster_dashboard_region_labeu-west-1-dev-2": "LABEU (France - Croix) (DEV-2)"
