+++ packages/manager/apps/dedicated/client/app/dedicated/dedicated-server/clusters/cluster/dashboard/index.js
@@ -5,12 +5,10 @@ import '@uirouter/angularjs';
 import 'angular-translate';
 import '@ovh-ux/ui-kit';
 
-import {
-  clusterDashboard,
-  clusterGeneralInformation,
-} from '@ovh-ux/manager-bm-cluster-components';
 import ovhManagerBillingComponents from '@ovh-ux/manager-billing-components';
 import { serverSupport } from '@ovh-ux/manager-bm-server-components';
+import clusterGeneralInformation from './general-information';
+import component from './dashboard.component';
 import routing from './dashboard.routing';
 
 const moduleName = 'ovhManagerDedicatedClusterDashboard';
@@ -22,9 +20,9 @@ angular
     'ui.router',
     ovhManagerBillingComponents,
     serverSupport,
-    clusterDashboard,
     clusterGeneralInformation,
   ])
+  .component('clusterDashboard', component)
   .config(routing)
   .run(/* @ngTranslationsInject:json ../translations */);
