--- packages/manager/modules/bm-cluster-components/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@ovh-ux/manager-bm-cluster-components",
-  "version": "1.0.0",
-  "private": true,
-  "description": "OVHcloud BmClusterComponents product",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ovh/manager.git",
-    "directory": "packages/manager/modules/bm-cluster-components"
-  },
-  "license": "BSD-3-Clause",
-  "author": "OVH SAS",
-  "main": "./src/index.js",
-  "dependencies": {
-    "@ovh-ux/manager-config": "^7.1.1",
-    "angular-ui-validate": "angular-ui/ui-validate#~1.2.2",
-    "bootstrap4": "twbs/bootstrap#v4.6.2",
-    "ipaddr.js": "^1.6.0",
-    "lodash": "^4.17.11",
-    "moment": "^2.24.0",
-    "punycode": "^1.4.1",
-    "urijs": "^1.19.2"
-  },
-  "peerDependencies": {
-    "@ovh-ux/manager-core": "^12.0.0 || ^13.0.0",
-    "@ovh-ux/manager-ng-layout-helpers": "^2.0.0",
-    "@ovh-ux/ng-at-internet": "^5.9.1",
-    "@ovh-ux/ng-ovh-feature-flipping": "^1.0.5",
-    "@ovh-ux/ng-ovh-utils": "^14.0.17",
-    "@ovh-ux/ng-translate-async-loader": "^2.1.4",
-    "@ovh-ux/ui-kit": "^6.5.3",
-    "@uirouter/angularjs": "^1.0.23",
-    "angular": "^1.7.5",
-    "angular-translate": "^2.18.1",
-    "oclazyload": "^1.1.0"
-  }
-}
