--- packages/manager/modules/bm-cluster-components/src/cluster-dashboard/index.js
@@ -1,15 +1,0 @@
-import angular from 'angular';
-
-import '@ovh-ux/ui-kit';
-import 'angular-translate';
-
-import component from './dashboard.component';
-
-const moduleName = 'ovhManagerBmClusterComponentsClusterDashboard';
-
-angular
-  .module(moduleName, ['oui', 'pascalprecht.translate'])
-  .component('clusterDashboard', component)
-  .run(/* @ngTranslationsInject:json ./translations */);
-
-export default moduleName;
