--- packages/manager/modules/bm-cluster-components/src/index.js
@@ -1,3 +1,0 @@
-export { default as clusterMainPage } from './cluster';
-export { default as clusterDashboard } from './cluster-dashboard';
-export { default as clusterGeneralInformation } from './general-information';
